requirements.txt
Django==4.2.7
django-debug-toolbar==4.3.0
psycopg==3.1.17
psycopg-binary==3.1.17


djangorestframework~=3.14.0